Create EpEverSolarMonitorMQTT2iobroker.ino
I added the integration via mqtt to iobroker with user and password . This was tested and is successfully working with a Epever tracer 3210AN via Wemos D1 mini and MAX485. I took the 5V from the USB port and powered the Wemos D1 mini and the MAX485 with 5V . Important, not to use the Epever RJ45 PIN 1 !!!
I only used PIN for A / B / GND .
	 
Credits to all people involved above.